# server/openapi.yaml
openapi: 3.1.0
info:
  title: Video Analysis Runtime API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:18765
security:
  - bearerAuth: []
paths:
...
  /v1/assets:
    post:
      summary: Upload a raw media body or import an HTTP(S) URL with JSON
      parameters:
        - in: header
          name: X-Filename
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, contentEncoding: binary }
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
                filename: { type: string }
      responses:
        '201': { description: Asset created }
...
  /v1/jobs:
    post:
      summary: Submit an atomic operation or sequential pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
      responses:
        '202': { description: Job accepted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    JobRequest:
      type: object
      required: [operation]
      properties:
        operation:
          enum:
            - media.probe
            - media.extract_audio
            - media.extract_frames
            - asr.transcribe
            - llm.generate
            - video.analyze
            - connector.douyin.import
            - pipeline.run
        input: { type: object }
        stages:
          type: array
          items:
            type: object
            required: [id, type]
            properties:
              id: { type: string }
              type: { type: string }
              input: { type: object }